+++ R/lockfile-read.R
@@ -39,12 +39,47 @@ renv_lockfile_read_finish <- function(data) {
 
 }
 
+renv_lockfile_read_preflight <- function(contents) {
+
+  # check for merge conflict markers
+  starts <- grep("^[<]+", contents)
+  ends   <- grep("^[>]+", contents)
+
+  hasconflicts <-
+    length(starts) &&
+    length(ends) &&
+    length(starts) == length(ends)
+
+  if (hasconflicts) {
+
+    parts <- .mapply(function(start, end) {
+      c(contents[start:end], "")
+    }, list(starts, ends), NULL)
+
+    all <- unlist(parts, recursive = TRUE, use.names = FALSE)
+
+    renv_pretty_print(
+      values    = head(all, n = -1L),
+      preamble  = "The lockfile contains one or more merge conflict markers:",
+      postamble = "You will need to resolve these merge conflicts before the file can be read.",
+      emitter   = message,
+      wrap      = FALSE
+    )
+
+    stop("lockfile contains merge conflict markers; cannot proceed", call. = FALSE)
+
+  }
+
+}
+
 renv_lockfile_read <- function(file = NULL, text = NULL) {
 
   contents <- if (is.null(file))
     unlist(strsplit(text, "\n", fixed = TRUE))
   else
     readLines(file, warn = FALSE, encoding = "UTF-8")
+
+  renv_lockfile_read_preflight(contents)
 
   # try reading as JSON
   json <- catch(renv_json_read(text = contents))